ShortestPaths Pretty Much Done
Shortest Paths for Distance and Price work, small bug when there is no path between source and destination, and source is not the first element in the vector. Doesn't crash, but prints path with random number as cost. .txt has had a new city added ("No") to test this, no connections